requirements.txt
aiohttp>=2.31.0
requests>=2.31.0
pillow>=10.0.0
pydantic>=2.12.3
asyncio>=3.4.3
pyyaml>=6.0.1
GitPython>=3.1.40
PyGithub>=2.1.1
pyfiglet>=1.0.4
python-dotenv>=1.2.1